done prototype

+++ backend/src/main/java/nosql/zkh/backend/dto/GeotagDto.java
@@ -1,10 +1,10 @@
 package nosql.zkh.backend.dto;
 
 public class GeotagDto {
-    private Float latitude;
-    private Float longitude;
+    private Double latitude;
+    private Double longitude;
 
-    public GeotagDto(Float latitude, Float longitude) {
+    public GeotagDto(Double latitude, Double longitude) {
         this.latitude = latitude;
         this.longitude = longitude;
     }
@@ -12,19 +12,19 @@ public class GeotagDto {
     public GeotagDto() {
     }
 
-    public Float getLatitude() {
+    public Double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(Float latitude) {
+    public void setLatitude(Double latitude) {
         this.latitude = latitude;
     }
 
-    public Float getLongitude() {
+    public Double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(Float longitude) {
+    public void setLongitude(Double longitude) {
         this.longitude = longitude;
     }
 }
