+++ backend/src/main/java/nosql/zkh/backend/model/Appeal.java
@@ -21,9 +21,9 @@ public class Appeal {
 
     private String address;
 
-    private Float longitude;
+    private Double longitude;
 
-    private Float latitude;
+    private Double latitude;
 
     private String type;
 
@@ -38,7 +38,7 @@ public class Appeal {
     @Relationship(type = "WorksOn", direction = Relationship.Direction.INCOMING)
     public List<Worker> workerList;
 
-    public Appeal(Long id, String title, String description, String status, String feedback, String address, Float longitude, Float latitude, String type, LocalDateTime date) {
+    public Appeal(Long id, String title, String description, String status, String feedback, String address, Double longitude, Double latitude, String type, LocalDateTime date) {
         this.id = id;
         this.title = title;
         this.description = description;
@@ -95,19 +95,19 @@ public class Appeal {
         this.address = address;
     }
 
-    public Float getLongitude() {
+    public Double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(Float longitude) {
+    public void setLongitude(Double longitude) {
         this.longitude = longitude;
     }
 
-    public Float getLatitude() {
+    public Double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(Float latitude) {
+    public void setLatitude(Double latitude) {
         this.latitude = latitude;
     }
